feat: add partition_info field in system.public.tables
